updated to use proper naming convention
